+++ packages/workers/src/data/entities/job-audit.entity.ts
@@ -10,6 +10,8 @@ export class JobAuditEntity extends BaseEntity {
 
   workerId!: string;
 
+  emitter!: string;
+
   currentAttempt!: number;
 
   startedAt?: Date;