+++ packages/database/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.0
+
+### Major Changes
+
+- fix: update job audit emitter to be required
 ## 1.17.1
 
 ### Patch Changes