+++ packages/messaging/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.3.0
+
+### Minor Changes
+
+- feat: add job envelope type
+- fix: update job audit emitter to be required
+- Updated dependencies []:
+  - @volontariapp/database@2.0.0
 ## 1.2.3
 
 ### Patch Changes